[Add] テストファイルを追加して、テスト可能にした
