+++ ios-training-Ryosuke/Controller/WeatherViewController.swift
@@ -11,9 +11,9 @@ import Combine
 final class WeatherViewController: UIViewController {
 
     private var cancellables: Set<AnyCancellable> = []
-    @IBOutlet private weak var weatherImage: UIImageView!
-    @IBOutlet private weak var maxTemperatureLabel: UILabel!
-    @IBOutlet private weak var minTemperatureLabel: UILabel!
+    @IBOutlet weak var weatherImage: UIImageView!
+    @IBOutlet weak var maxTemperatureLabel: UILabel!
+    @IBOutlet weak var minTemperatureLabel: UILabel!
     private var weatherModel: WeatherModel
     
     init?(coder: NSCoder, weatherModel: WeatherModel) {
